# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crew Intelligence Platform — AI Services API
  description: |
    REST API for the CIP Python microservices layer.
    Provides Excel parsing, legality engine, AI assistant (Najm), smart ranking, and auto-bid services.
    
    **Authentication:** All endpoints require a valid Firebase ID token passed as `Authorization: Bearer <token>`.
    Service-to-service calls use a signed service account JWT.
    
    **Base URL (production):** `https://cip-python-xxxx.run.app`
    **Base URL (development):** `http://localhost:8080`
  version: 1.0.0
  contact:
    name: CIP Support
    email: support@cip.app
servers:
  - url: https://cip-python-xxxx.run.app
    description: Production (Google Cloud Run — me-central1)
  - url: http://localhost:8080
    description: Local development
security:
  - FirebaseAuth: []
...
paths:
...
  /v1/roster-sync/connections:
    post:
      tags: [Roster Sync]
      summary: Register a connection (status tracking only — NEVER credentials)
      description: |
        Credentials live exclusively in device secure storage. This endpoint
        (like every roster-sync endpoint) rejects any payload containing a
        credential-shaped key with 422 (`assert_no_credentials`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider_id]
              properties:
                provider_id: { type: string, example: ics_feed }
                user_id: { type: string, description: Service callers only. }
                client_meta: { type: object }
      responses:
        "200":
          description: Connection state
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [connected, awaiting_official_integration] }
                  note: { type: string }
        "422": { description: Credential-shaped key detected in payload }
  /v1/roster-sync/connections/{provider_id}:
    delete:
      tags: [Roster Sync]
      summary: Disconnect (imported rosters are kept)
      parameters:
        - { name: provider_id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Disconnected; device performs the local credential wipe }
  /v1/roster-sync/connections/{provider_id}/sync-now:
    post:
      tags: [Roster Sync]
      summary: Manual sync trigger
      description: |
        server_orchestrated providers sync here; client_orchestrated
        providers answer `client_sync_required` and the device performs the
        fetch + import.
      parameters:
        - { name: provider_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Action taken
          content:
            application/json:
              schema:
                type: object
                properties:
                  action: { type: string, enum: [server_synced, client_sync_required, unavailable] }
                  detail: { type: string }
  /v1/roster-sync/import:
    post:
      tags: [Roster Sync]
      summary: Import a fetched roster (dedup + versioning + engine fan-out)
      description: |
        Identity is pinned to the caller's token. Duplicate rosters
        (identical checksum) short-circuit without writes; changed rosters
        produce a leg-level diff, supersede the previous ACTIVE line
        (kept as history), and trigger the engine fan-out. Failures leave
        the previous roster untouched (422 with a meaningful message).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RosterImportRequest'
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RosterImportResponse'
        "409": { description: Provider awaiting official integration }
        "422": { description: Parse failure or credential-shaped key — previous roster kept }
...
components:
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
  schemas:
...
    RosterImportRequest:
      type: object
      required: [provider_id, period, year, payload_kind, payload]
      properties:
        provider_id: { type: string }
        user_id: { type: string, description: Service callers only. }
        period: { type: string, example: JUL-2026 }
        year: { type: integer, example: 2026 }
        payload_kind: { type: string, enum: [ics, normalized] }
        payload:
          description: Raw ICS text, or a normalized roster object. Never credentials.
    RosterImportResponse:
      type: object
      properties:
        result: { type: string, enum: [imported, duplicate, failed] }
        line_id: { type: string, nullable: true }
        version: { type: integer, nullable: true }
        imported_flights: { type: integer }
        checksum: { type: string, nullable: true }
        diff:
          type: object
          properties:
            added: { type: integer }
            removed: { type: integer }
            changed: { type: integer }
        engines:
          type: array
          items:
            type: object
            properties:
              engine: { type: string, example: salary_engine }
              status: { type: string, enum: [triggered, queued, on_demand, error] }
              detail: { type: string }
...